+++ fragnet-search/CHANGES.md
@@ -19,3 +19,6 @@
 
 ### 26-AUG-2019
 * Initial MCS implementation for groups giving the `refmolAtomsMissing` property
+
+### 29-AUG-2019
+* better handling of duplicate paths when generating groups